controller/requirements.txt
cbor==1.0.0
coppeliasim_zmqremoteapi_client==2.0.4
et_xmlfile==2.0.0
numpy==2.3.1
opencv-python==4.13.0.92
openpyxl==3.1.5
paho-mqtt==2.1.0
pandas==3.0.2
python-dateutil==2.9.0.post0
pyzmq==27.0.0
six==1.17.0
